Warning fixes CompteShaderManager

+++ DoremiEngine/Graphic/Source/Manager/ComputeShaderManagerImpl.cpp
@@ -214,8 +214,8 @@ namespace DoremiEngine
             shaderManager.SetActiveComputeShader(m_blurHoriShader);
             DirectX::XMFLOAT2 t_screenRes = m_graphicContext.m_graphicModule->GetSubModuleManager().GetDirectXManager().GetScreenResolution();
 
-            UINT x = ceil(t_screenRes.x / (FLOAT)(256 * 2));
-            UINT y = t_screenRes.y / 2;
+            UINT x = static_cast<UINT>(ceil(t_screenRes.x / (FLOAT)(256 * 2)));
+            UINT y = static_cast<UINT>(t_screenRes.y / 2);
 
             m_deviceContext->Dispatch(x, y, 1);
         }
@@ -226,8 +226,8 @@ namespace DoremiEngine
             shaderManager.SetActiveComputeShader(m_blurVertShader);
             DirectX::XMFLOAT2 t_screenRes = m_graphicContext.m_graphicModule->GetSubModuleManager().GetDirectXManager().GetScreenResolution();
 
-            UINT x = t_screenRes.x / 2;
-            UINT y = ceil(t_screenRes.y / (FLOAT)(256 * 2));
+            UINT x = static_cast<UINT>(t_screenRes.x / 2);
+            UINT y = static_cast<UINT>(ceil(t_screenRes.y / (FLOAT)(256 * 2)));
 
             m_deviceContext->Dispatch(x, y, 1);
         }
